refactor(pipeline/scrape): split ScrapePhaseActions into ScrapePhase/ leaf modules (Phase 8.5b commit 4)
Extract the 986-LoC ScrapePhaseActions god-file into 5 cohesive
leaf modules under Mediator/Scrape/ScrapePhase/ + a barrel. The
parent file collapses to a 271-LoC composer that re-exports the
public surface and keeps the four executeXxx orchestrators
(forensic-pre, matrix-loop, validate-results, stamp-accounts +
internal direct-discovery). All five leaf modules respect the
C6-target 10-eff per-function cap; six over-cap functions were
refactored during extraction:

- buildAccountIdentities  13 -> 8  (Identity.ts)        map(... ): tuple + new Map(entries)
- recordToIdentity        11 -> 10 (Identity.ts)        lookupQueryIdValue + lookupBankAccountIdValue helpers
- discoverBalanceFetchTemplate 13 -> 8 (BalanceTemplate.ts)  tryFinderChain + buildKeyedCandidates helpers
- decideEmptyGate         16 -> 8  (EmptyDetection.ts)  IGateOutcomeArgs bundle + failOnEmptyAllAccounts
- warnZeroAmounts         12 -> 9  (EmptyDetection.ts)  shouldWarnAllZero predicate
- (tryFinderChain itself  -> 8  via buildKeyedCandidates extraction)

Module layout (each <=10 eff per fn, well below =12A grandfather):

- Diag.ts            43 raw /  19 eff /  2 fns  - LOG + buildPreDiag + maybeForensicPrime
- PreDiscovery.ts   147 raw /  67 eff /  6 fns  - 6 readers (pre-discovered accounts + txn + dashboard + billing + dedup + date-window) + EMPTY_* + IPreDiscoveredAccounts
- Identity.ts       117 raw /  53 eff /  6 fns  - identity triples + BANK_ACCOUNT_ID_FIELDS
- BalanceTemplate.ts 326 raw / 143 eff / 20 fns - POST/GET-query/GET-path/Bulk finders + parseQueryRecord + EMPTY_JSON_PARSE + EMPTY_BALANCE_TEMPLATE
- EmptyDetection.ts 207 raw /  95 eff / 10 fns  - real-empty gating + zero-amount audit + storage-safe + 4 IArgs bundles
- index.ts           33 raw /  26 eff /  0 fns  - barrel re-exports

ScrapePhaseActions.ts remains under =12A grandfather for now
(four composer fns still over cap: executeForensicPre 4 eff,
executeMatrixLoop 4 eff, collectStorageSafe 11 eff,
executeValidateResults 11 eff, executeStampAccounts 20 eff,
executeDirectDiscovery 73 eff). C5 will move these into
ScrapePhase/Actions.ts and refactor each to <=10 eff. C6 then
drops the =12A grandfather entirely.

Public surface preserved through re-exports:
- 2 type aliases: EMPTY_TXN_ENDPOINT, EMPTY_TXN_HARVEST
- 5 readers consumed by FrozenScrapeAction.ts:
  readBillingCycleCatalog, readDashboardTxnHarvest,
  readDateWindowParams, readDedupKeyFields, readPreDiscoveredTxn
- 4 executeXxx consumed by ScrapeMediator:
  executeForensicPre, executeMatrixLoop, executeStampAccounts,
  executeValidateResults

Verification:
- lib/index.d.ts SHA256: 221E161123230F013CB4687CE4CD51109DA5C7D579552DE5328190AD98AAB055 (byte-identical to Phase 8 baseline)
- lib/index.cjs: 877,422 bytes (+842 vs C3 due to new sub-module wrappers; +2,687 vs main)
- Madge cycles: 18 (no new vs main baseline 18)
- Jest Scrape cluster: 179/179 pass (16 suites: ScrapePhase/Actions/Branches/Wave5/Factory + FrozenScrapeAction[Wave5] + ScrapeReplayAction[Wave5] + GenericScrapeReplay)
- All 6 ScrapePhase/*.ts files: 0 functions over 10-eff cap, 0 ESLint errors, 0 TSC errors
